Add files via upload
I created a books website that allows you to add books to the basket and favorites box.
Because I did not use the databases on the site, I manipulated Local Storage to add the books to the cart.
On the basket page, the user can increase the number of books or remove books from the basket, and the total of books will change automatically. There is also a feature of using the discount code. When the correct discount code is used, you can deduct the amount from the total and you can remove the discount code.